requirements.txt
requests
beautifulsoup4
trafilatura
lxml
tqdm
loguru
underthesea==1.3.3
onnxruntime==1.22.1
qdrant-client==1.16.1
sympy==1.14.0
torch==2.9.1
python-dotenv==1.2.1
pandas
rank-bm25==0.2.2
numpy==1.26.4